Split tmux screen routing helpers
