+++ scripts/file-size-hygiene-baseline.txt
@@ -9,7 +9,7 @@
 207337 app/src/main/java/com/pocketshell/app/composer/PromptComposerViewModel.kt
 139372 app/src/main/java/com/pocketshell/app/projects/FolderListScreen.kt
 180197 app/src/main/java/com/pocketshell/app/projects/FolderListViewModel.kt
-303487 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt
+271816 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionScreen.kt
 994135 app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
 290559 app/src/test/java/com/pocketshell/app/composer/PromptComposerViewModelTest.kt
 167942 app/src/test/java/com/pocketshell/app/session/AgentConversationRepositoryTest.kt
